feat: force checkout

+++ packages/cli/src/projects/clean.ts
@@ -3,19 +3,24 @@ import { Workspace } from '@openfn/project';
 import { rimraf } from 'rimraf';
 
 import { build, ensure } from '../util/command-builders';
+import { handler as checkout } from './checkout';
 import type { Logger } from '../util/logger';
 import * as o from '../options';
 import * as po from './options';
 
 import type { Opts } from './options';
 
-export type CleanOptions = Pick<Opts, 'command' | 'workspace' | 'log'>;
+export type CleanOptions = Pick<
+  Opts,
+  'command' | 'workspace' | 'log' | 'confirm' | 'force'
+>;
 
-const options = [o.log, po.workspace];
+const options = [o.log, o.confirm, o.force, po.workspace];
 
 const command: yargs.CommandModule = {
   command: 'clean',
-  describe: 'Delete the workflows folder for the currently active project',
+  describe:
+    'Delete the workflows folder and re-checkout the currently active project',
   handler: ensure('project-clean', options),
   builder: (yargs) => build(options, yargs),
 };
@@ -26,6 +31,26 @@ export const handler = async (options: CleanOptions, logger: Logger) => {
   const workspacePath = options.workspace ?? process.cwd();
   const workspace = new Workspace(workspacePath, logger);
 
+  const skip = options.force || options.confirm === false;
+  const doIt = await logger.confirm(
+    `This will delete all files in ${workspace.workflowsPath}. Do you want to proceed?`,
+    skip
+  );
+  if (!doIt) {
+    return;
+  }
+
   await rimraf(workspace.workflowsPath);
-  logger.success(`Removed workflows directory at ${workspace.workflowsPath}`);
+
+  const activeProject = workspace.activeProject;
+  if (!activeProject) {
+    throw new Error(
+      'No active project found in workspace. Run `project pull` first.'
+    );
+  }
+
+  await checkout(
+    { ...options, project: String(activeProject.uuid), force: true },
+    logger
+  );
 };
